merge: remember conflict labels
When recreating merge conflicts with "git checkout -m <path>" the
original conflict labels are lost. For commands like "git merge" and
"git cherry-pick" we could use the presence of the related root
ref (MERGE_HEAD and CHERRY_PICK_HEAD respectively) to recreate the
labels. However if the conflicts are from "git stash pop" or "git
checkout -m <branch>" then there is no ref to deduce the labels from. To
ensure the labels are always available the merge machinery is updated to
record the labels .git/MERGE_LABELS when it updates the worktree and
there are conflicts. The labels are then read from the file by "git
checkout -m <path>" when recreating the conflicts.

There is a subtlety with "git checkout -m <branch>" in that it resets the
index to remove the conflict entries but they are saved in merge undo so
we want to preserve the MERGE_LABELS file when switching branches if
there were conflicts.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>