fixup! git-std-lib: introduce git standard library
Rather than re-using LIB_OBJS when building git-std-lib.a use a separate
variable to hold the list of sources and build git using git-std-lib.a
so we don't have to worry about having the same source file in two
different libraries. Also move the stub objects into their own
library. This enables git to be built using git-std-lib.a with

    make git

which will link git against git-std-lib.a built with gettext and tracing
support. To use git-std-lib.a in other programs compile it and
git-stub-lib.a with

    make NO_GETTEXT=YesPlease git-std-lib.a git-stub-lib.a

and link your program against both libraries.

The GIT_STD_LIB define is removed in favor of more stubbing to avoid
the complications of conditional compilation.

The Documentation is also tweaked to allow the html file to be built.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>