fixup! git-std-lib: add test file to call git-std-lib.a functions
The interesting changes here are in the Makefile which now compiles and
links stdlib-test with our usual CFLAGS and LDFLAGS and also links
against git-stub-lib.a as well as git-std-lib.a. The remaining changes
are to silence compiler warnings. Hopefully this whole thing will be
replaced by proper unit tests in the future.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>