checkout: generalize `writeout_error`
The `writeout_error` parameter is used to propagate an error up the
call-chain if the files in the work-tree cannot be updated. This error
is propagated independently of the return value of the function. The
next commit will want to propagate other errors in this way so prepare
for that by renaming the variable to `error_flags`

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>