add -p: mark split hunks as undecided
When a hunk is split each of the new hunks inherits whether it is
selected or not from the original hunk. If a selected hunk is split
all of the new hunks are selected and the user is only prompted with
the first of the split hunks, they are not asked whether or not they
want to select the rest of the new hunks. If they want to deselect
one of the other new hunks they have to navigate back to it to do
that. This is unfortunate as the user is presumably splitting the
original hunk because they only want to select some sub-set of it.

Instead mark all the new hunks as "undecided" so that we prompt the
user to decide whether to select them or not. In the case where the
user only wants to change the selection of the first of the split
hunks they will now have to do more work re-selecting the remaining
split hunks. However changing the selection of any of the other newly
created hunks is now much simpler as the user no-longer has to navigate
back to them before being able to change their selected state. Due
to concerns that users may be relying on the current behaviour [1]
this change is guarded by WITH_BREAKING_CHANGES.

[1] https://lore.kernel.org/git/xmqqjz9b6xr1.fsf@gitster.g

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>