add-patch: update hunk splitability after editing
When the users edits a hunk if they change deletion lines to context
lines or vice versa then the number of hunks that the edited hunk can be
split into may differ from the unedited hunk and so we need to update
hunk->splittable_into. In practice users are unlikely to hit this bug as
it is doubtful that a user who has edited a hunk will split it
afterwards.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>