midx docs: clarify tie breaking
Clarify what happens when an object exists in more than one pack, but
not in the preferred pack. If the user does not pass a preferred pack
then the pack with the lowest mtime is chosen as the preferred pack. For
objects that are not in the preferred pack the pack with the highest
mtime is used. "git multi-pack-index repack" relies on this behavior. If
ties were resolved in favor of the oldest pack as the current
documentation suggests the multi-pack index would not reference any of
the objects in the pack created by "git multi-pack-index repack".

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>