breaking-changes: switch default branch to main
Since ... "git-init.adoc" has advertized that the default name of the
initial branch may change in the future. The breaking changes document
says that we'll update it to "main" in git 3.0. This patch implements
that. The name "main" is chosen to match the default used by the big
git forge web sites. The code changes are quite straight forward.

With the expection of the "git-init.adoc" the documentation is left
unchanged. I had hoped to parameterize the name of the default branch by
using an asciidoc attribute but unfortunately attribute expansion is
inhibited by backticks and we use backticks to mark up ref names so that
idea does not work. As the changes to git-init.adoc show inserting
ifdef's around each instance of the branch name "master" is cumbersome
and makes the documentation sources harder to read.

Apart from "git-init.adoc" I think the only other place where master is
used as the name of the initial branch rather than as an example of a
branch name is in "gitcore-tutorial.adoc". The name appears a lot in the
tutorial so updating it with ifdef's is not really practical. We can
update that document in the 3.0 release cycle. The other documentation
where master is used as an example branch name can be gradually
converted over time.

Now that none of the tests rely on
GIT_TEST_DEFAULT_INITIAL_BRANCH_NAME=master we can remove support for
that as well.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>