commit: print advice when core.commentString=auto
Support for this config setting was deprecated in the last commit so
print some advice to help the user update their config settings. The
advice message explains that the setting is deprecated and will be
removed in future. It also shows the commands that the user needs to run
to either unset core.commentChar and core.commentString completely or to
change the current setting to a fixed comment string.

In order to generate this advice we need to parse the config with a
callback that records each file where either of the keys is set and
whether a key occurs more that once in a given file. This lets us
generate the list of commands to remove all the keys and also tells us
which key the "auto" setting comes from. The hard coding of some
filenames in add_config_scope_arg() is unfortunate but as this temporary
code that will be removed when Git 3.0 is released I decided it wasn't
worth adding functions to get the name of the local and worktree config
files.

As we want the user to update their config we do not provide a way for
this advice to be disabled other than changing the value of
core.commentChar or core.commentString.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>