t0018: switch default branch name to main
These tests use "trunk" as the default branch name but the exact
name of the branch is incidental to testing if the advice message
includes it. Git 3.0 will change the default branch name to "main"
and remove support for GIT_TEST_DEFAULT_INITIAL_BRANCH_NAME so change
these test to use "main" instead of "topic".

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>